Fix reinforcements
It can be a bit hard to test because you can only do one per
cluster, so if there are only 2 stuck in australia that doesn't
mean he should move it. However in the late game it's very clear:
long attacks and at the end the army is moved to the front.